Added end chase vision collider, and updated vision cone script
Added a new vision cone to test enemy, which is the cone that you should need to escape to make the enemy stop chasing. Additionally, updated vision cone script to work better with other base radiuses, as it currently didn't, but it's likely still a bit broken, but the adjustment variable can compensate.
Also changed camera scale.